Fix path issues as pointed out by @DeliciousNukaCola.
os.chdir() can't change to an empty string value, which is what was
being returned under certain circumstances.

Instead, training program has been updated make a note of the location
of the module at runtime, used to figure out where its support files are
located.

Also, it now writes its output to the current directory rather than the
module's directory, which is probably what you wanted when you ran it
from outside of the project directory in the first place.